Move ShippableItemUI and ShippableItemsUI to view model
