+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/WooShippingLabelCreationViewModel.kt
@@ -857,4 +857,24 @@ data class WooShippingAddresses(
     }
 }
 
-fun ShippableItemsUI.totalItemQuantity() = shippableItems.sumByFloat { it.quantity }.toInt()
+@Parcelize
+data class ShippableItemUI(
+    val itemId: Long,
+    val productId: Long,
+    val title: String,
+    val formattedSize: String,
+    val formattedWeight: String,
+    val formattedPrice: String,
+    val quantity: Float,
+    val imageUrl: String? = null
+) : Parcelable
+
+@Parcelize
+data class ShippableItemsUI(
+    val shippableItems: List<ShippableItemUI>,
+    val formattedTotalWeight: String,
+    val formattedTotalPrice: String
+) : Parcelable {
+    val totalItemQuantity
+        get() = shippableItems.sumByFloat { it.quantity }.toInt()
+}
